+++ src/game/TileEngine/Render_Dirty.h
@@ -35,8 +35,8 @@ struct VIDEO_OVERLAY
 	SGPFont          uiFontID;
 	INT16            sX;
 	INT16            sY;
-	UINT8            ubFontBack;
-	UINT8            ubFontFore;
+	UINT32           ubFontBack;
+	UINT32           ubFontFore;
 	wchar_t          zText[200];
 	SGPVSurface*     uiDestBuff;
 	OVERLAY_CALLBACK BltCallback;
@@ -78,7 +78,7 @@ void GPrintInvalidateF(INT16 x, INT16 y, wchar_t const* fmt, ...);
 
 // VIDEO OVERLAY STUFF
 VIDEO_OVERLAY* RegisterVideoOverlay(OVERLAY_CALLBACK callback, INT16 x, INT16 y, INT16 w, INT16 h);
-VIDEO_OVERLAY* RegisterVideoOverlay(OVERLAY_CALLBACK callback, INT16 x, INT16 y, SGPFont font, UINT8 foreground, UINT8 background, wchar_t const* text);
+VIDEO_OVERLAY* RegisterVideoOverlay(OVERLAY_CALLBACK callback, INT16 x, INT16 y, SGPFont font, UINT32 foreground, UINT32 background, wchar_t const* text);
 void ExecuteVideoOverlays(void);
 void SaveVideoOverlaysArea(SGPVSurface* src);
